# new permutedims without unrolling permutation orders