--- a/src/host/linalg.jl
+++ b/src/host/linalg.jl
@@ -180,19 +180,23 @@ LinearAlgebra.lmul!(a::Number, B::AbstractGPUArray) = generic_lmul!(a, B)
 
 
 ## permutedims
-
-# TODO: implementation without the memory copy
 LinearAlgebra.permutedims!(dest::AbstractGPUArray, src::AbstractGPUArray, perm) =
     permutedims!(dest, src, Tuple(perm))
 
 function LinearAlgebra.permutedims!(dest::AbstractGPUArray, src::AbstractGPUArray,
                                     perm::NTuple{N}) where N
     Base.checkdims_perm(dest, src, perm)
+
+    # get the new strides of destination tensor
     dest_strides = ntuple(k->k==1 ? 1 : prod(i->size(dest, i), 1:k-1), N)
     dest_strides_perm = ntuple(i->dest_strides[findfirst(==(i), perm)], N)
+
     function permutedims_kernel(ctx, dest, src, dest_strides_perm)
+        # find the cartesian index in source tensor
         LI = @linearidx src
         I = @inbounds CartesianIndices(src)[LI]
+
+        # the corresponding linear index in the destination tensor
         dest_index = map_index(I.I, dest_strides_perm)
         @inbounds dest[dest_index] = src[LI]
         return