--- a/test/testsuite/linalg.jl
+++ b/test/testsuite/linalg.jl
@@ -16,7 +16,9 @@
         @test compare(x -> permutedims(x, (3, 1, 2)), AT, rand(Float32, 4, 5, 6))
         @test compare(x -> permutedims(x, [2,1,4,3]), AT, randn(ComplexF32,3,4,5,1))
         # high dimensional tensor
-        @test compare(x -> permutedims(x, 18:-1:1), AT, rand(Float32, 4, [2 for _ = 2:18]...))
+        @static if VERSION >= v"1.7"
+            @test compare(x -> permutedims(x, 18:-1:1), AT, rand(Float32, 4, [2 for _ = 2:18]...))
+        end
     end
 
     @testset "symmetric" begin